move api to reporter head
